requirements.txt
# Core Flask dependencies
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0

# AI and ML dependencies
google-generativeai==0.8.3
sentence-transformers==2.2.2
chromadb==0.4.22

# Document processing
PyPDF2==3.0.1
python-docx==1.1.0

# Additional utilities
Werkzeug==3.0.1
requests==2.31.0